refactor(builtins): Move builtins in builtins/ dir
