+++ lua/core/builtins/init.lua
@@ -1,27 +1,27 @@
 local M = {}
 
 local builtins = {
-  "core.which-key",
-  "core.gitsigns",
-  "core.cmp",
-  "core.dashboard",
-  "core.dap",
-  "core.terminal",
-  "core.telescope",
-  "core.treesitter",
-  "core.nvimtree",
-  "core.project",
-  "core.bufferline",
-  "core.autopairs",
-  "core.comment",
-  "core.lspinstall",
-  "core.lualine",
+  "core.builtins.which-key",
+  "core.builtins.gitsigns",
+  "core.builtins.cmp",
+  "core.builtins.dashboard",
+  "core.builtins.dap",
+  "core.builtins.terminal",
+  "core.builtins.telescope",
+  "core.builtins.treesitter",
+  "core.builtins.nvimtree",
+  "core.builtins.project",
+  "core.builtins.bufferline",
+  "core.builtins.autopairs",
+  "core.builtins.comment",
+  "core.builtins.lspinstall",
+  "core.builtins.lualine",
 }
 
 function M.setup(config)
   for _, builtin_path in ipairs(builtins) do
     local builtin = require(builtin_path)
-    local config_path = builtin_path:gsub("core%.", ""):gsub("-", "_")
+    local config_path = builtin_path:gsub("core%.builtins%.", ""):gsub("-", "_")
     builtin:setup(config:sub(config_path))
   end
 end
